requirements.txt
langchain
python-dotenv
google-generativeai
Flask
